Finally I managed to compile everything into the first post. Well this is not quite correct unfortunatly I hit the limit of characters I can put into one post so I used my second post in that thread to give all the information.
As base I used Keygen's all in one pack and I added all the text files you can
find in the patch and some other additions here is the last part of the stuff
I indended to put into the first post:

..\ctp2_code\ui\interface\dipwirzard.cpp
..\ctp2_code\ui\interface\dipwirzard.h
..\ctp2_data\XXXXXX\uidata\layouts\dipwizard.ldl

Bug fix by Martin Gühmann: Added missing diplomacy photos to the diplomanager.


..\ctp2_code\ui\interface\spnewgameplayersscreen.cpp
..\ctp2_code\ui\interface\spnewgameplayersscreen.h
..\ctp2_data\XXXXXX\uidata\layouts\spnewgamepopups.ldl

Where XXXXXX stands for english, french, german, italian or spanish.

Interface Improvement by Martin Gühmann: Redesigned the sp new game player
screen, so that a new game can be now started with more then 8 players. 31
players is the maximum number of players that can be now used, don't expect
that your computer runs smoth with so many players. There is now also the
possibilty to set the maximum number of players that can occur in a game for
instance through revolts, maximum is also here 31 players. Another option is
there to select the player index that determines in the end the player's
color, unfortunatly this doesn't work, yet but this can be fixed when
something else is fixed from the 1.1/1.11 patch list.

Note that if you play with MaxPlayer higher than NumPlayers and you use a
version of Diplomod like in ApolytonPack then the game crashes if a new player
is added and the total number of players is higher then you originally started
with.


..\ctp2_data\XXXXXX\uidata\layouts\spnewgame.ldl

Where XXXXXX stands for english, french, german, italian or spanish.

Added disclaimers in text to the initial screen by Klaus Kaan.
Improved by Martin Gühmann to make it possible to localize all the strings on
that screen.

This file replaces the image drawn by Locutus and therefore this image is not
anymore part of the *.zip file. Well it is big enough.

..\ctp2_data\XXXXXX\gamedata\ldl_str.txt

Where XXXXXX stands for english or german.

Added new strings for disclaimers, version and the new sp new game players, to
allow easy localization of all the stuff.

Note that for other language versions we still need tranlators to create a
French, Italian and a Spanish version of that file.


.. \Scenarios\MagnificentSamurai\scen0000\default\gam
edata\scenario.slc

Fix by Martin Gühmann: When you return to Shichinen via the Star Road a fort
should be created in Shichinen. Unfortunatly they give to the event the wrong
argument the fort has the index 8 not 7. Using the TerrainImprovementDB
functions avoids the need to fix this again if the indices are shifted again.



I packed a *.zip file with the altered files in the fashion of the mods. So if
you just unzip it it into your default CTP2 folder and you installed the
ctp2_code in the default CTP2 folder it overwrites the original source files.
So if you altered some of these files already I recommend to back up them
before and add your alteration to these files. Or alternativly you could unzip
them to somewhere else, and alter the files from this *.zip. I recommend it
anyway so you can use this file structure to add your own files and provide
such *.zip's, too.