2004.11.27 Slic builtin update by Martin Gühmann:
..\ctp2_data\english\gamedata\civ_str.txt
..\ctp2_data\french\gamedata\civ_str.txt
..\ctp2_data\german\gamedata\civ_str.txt
..\ctp2_data\italian\gamedata\civ_str.txt

Updated to match latest civ additions. Note the Italian file only contains
untranslated strings in English. The according files for the Spanish and
Japanese version are missing. Latest English version based on a file by
Locutus. Latest French version based on a file by Tamerlin.

..\ctp2_data\german\gamedata\str_loc.txt
..\ctp2_data\italian\gamedata\str_loc.txt

Localized civ grammer for the new civs, note that only the file for the
German version contains real data, the Italian file was filled with dummy
data. A French version could also have filled with dummy data but I think
in this case we can wait for the file with the real data. And note such
a file is missing for the Spanish version. I don't know whether there is
such an file for the Japanese version.