Miscellaneous changes by Martin Gühmann:
..\ctp2_code\gs\armydata.cpp
..\ctp2_code\gs\armydata.h
Made sure that the bombard range code does compile, unfortunatly it still 
does not work.

..\ctp2_data\french\gamedata\ldl_str.txt
..\ctp2_data\french\gamedata\civ_str.txt
..\ctp2_data\french\gamedata\str_loc.txt
Tamerlin's latest files for the French version.

..\ctp2_data\german\gamedata\str_loc.txt
Fixed some wrong grammar concerning genetives.

..\ctp2_data\english\gamedata\Great_Library.txt
..\ctp2_data\german\gamedata\Great_Library.txt
..\ctp2_data\italian\gamedata\Great_Library.txt
..\ctp2_data\spanish\gamedata\Great_Library.txt
Added Tamerlin's proposal for the computer age to the Great Libary. 
Note that the entries for the Italian and the Spanish version remain 
untranslated.