Bug fixes and slic fixes and slic additions compiled by Martin Gühmann:
..\ctp2_code\ai\strategy\squads\squad_Strength.cpp
Replaced Calvitix version of the > operator by a cleaner version, however
this version of the > operator is still experimental and therefore it is
disabled as well. At least it allows AI sea activity, unfortunately it still
slows down the game.

..\ctp2_code\gs\gameobj\bldque.cpp
Fixed the bug that the player doesn't receive a wanring when another player
starts to build the first wonder in the database.

..\ctp2_code\gs\slic\SlicFunc.h
..\ctp2_code\gs\slic\slicfunc.cpp
..\ctp2_code\gs\slic\SlicEngine.cpp
- Added new slic functions:
- IsOnSameContinent: Checks whether two locations are on the same continent,
  by Solver.
- AddSlaves: Adds slaves to a city, this function was suggested by The Big
  Mc. It is the event based version. The error reported here for instance
  doesn't seem to be releated to the function itsself. It seems there is a
  problem in the Governor::AssignSpecialist function a problem. It seems
  that it ignores the fact that slaves cannot be assigned to other tasks
  than working in the field. Obviously the problem also occurs if the city
  has a lot of slaves regardless how they were added.
- CreateUnit function doesn't crash anymore if the given unit type isn't an
  invalid unit type.

..\ctp2_code\gs\slic\slicif.h
..\ctp2_code\gs\slic\slicif.cpp
..\ctp2_code\gs\slic\slic.y
Slic database access is now also possible if the record name was used as well
as symbol in the context. That prevented the gain of great leaders in CTC for
instance. As decribed here.

..\ctp2_code\gs\slic\slic.l
..\ctp2_code\gs\slic\sliccmd.y
..\ctp2_code\gs\slic\SlicFrame.cpp
..\ctp2_code\gs\slic\SlicSegment.cpp
Addition of power operator: ** This addition is based on files by tombom
posted here. The precidence of the operator isn't solved, yet. But it is time
to include the files, because slic.y and slicif.cpp also contain changes
concerning this subject.