Vision update and miscellaneous corrections by Martin Gühmann
..\ctp2_code\gs\gameobj\bldque.cpp
Fixed a bug that leaded to a crash in the first not reported wonder bug fix.

..\ctp2_code\gs\gameobj\UnitData.cpp
When a city is conquered the city is no more removed from the map and
inserted into the map again. That saves some time. The removement of the
vision of the old owner is postponed until the city data changed hand, so that
the old owner get a last look on "his" city, and everything on the map like
city and surrounding terrain is shown with the new owner. If you loose a city
you should know who was it. And you should also know which land the attacker
has stolen.

..\ctp2_code\tiledmap.cpp
..\ctp2_code\tiledraw.cpp
..\ctp2_code\ui\aui_ctp2\radarmap.cpp
If you own a tile it is displayed on the map as yours even if you didn't
visit that tile recently. If fog of war is of or god mode is on the borders of
civs with that you have no contact are displayed.

..\ctp2_code\gs\world\UnseenCell.cpp
The city owner is now taken from the city's CityData instead from the unit
itsself. This allows the correct display of an owner if you loose a city.

..\ctp2_code\ui\aui_ctp2\InfoBar.cpp
If you move the cursor over a city without actor in the UnseenCell, the game
doesn't crash anymore. The missing actor problem was caused by how the owner
change was handled during city capture. This problem is now fixed at the root,
but old savegames might contain this problem. Another problem with destroyed
civilsations and there cities was also fixed.

..\ctp2_data\default\aidata\advancelists.txt
..\ctp2_data\default\aidata\buildlistsequences.txt
..\ctp2_data\default\aidata\strategies.txt
..\ctp2_data\default\aidata\unitbuildlist.txt
Replaced BuildListSequenceElement by those of GoodMod that are based
themselfes of those from MedPack.
Replaced build lists sequences by those of GoodMod: The Capitalization and
Infrastructure elements are removed from the sequences.
Replaced the advance lists and unit build lists by those of GoodMod. IIRC
there are based on those of MyMod or ApolytonPack.
These changes improve the AI.

..\ctp2_data\french\gamedata\ldl_str.txt
City Manager is now translated. This is the file by Tamerlin from 2005.02.07.

..\ctp2_data\italian\gamedata\info_str.txt
..\ctp2_data\german\gamedata\info_str.txt
Some minor alterations like correction a grammatical case in the German
version of the wonder almost complete message.